+++ apps/cookbook/src/app/examples/label-example/label-example.component.html
@@ -1,13 +1,13 @@
 <div class="example-container">
   <h1>Label</h1>
 
-  <h2>In item</h2>
+  <h2>In Item</h2>
   <div class="example-frame">
-    <cookbook-label-example-item></cookbook-label-example-item>
+    <cookbook-label-example-direction></cookbook-label-example-direction>
   </div>
 
-  <h2>Label</h2>
+  <h2>Other label types</h2>
   <div class="example-frame">
-    <cookbook-label-form-field-input-label-example></cookbook-label-form-field-input-label-example>
+    <cookbook-form-field-input-label-example></cookbook-form-field-input-label-example>
   </div>
 </div>
