+++ apps/cookbook/src/app/examples/label-example/label-example.component.ts
@@ -1,18 +1,11 @@
 import { Component } from '@angular/core';
-import { LabelExampleItemComponent } from './examples/item';
-import { FormFieldTextareaLabelExampleComponent } from './examples/textarea/label';
-import { LabelFormFieldInputLabelExampleComponent } from './examples/label';
-import { RadioInFormFieldExampleComponent } from './examples/in-form-field';
+import { FormFieldInputLabelExampleComponent } from '../form-field-example/examples/input/label';
+import { LabelExampleDirectionComponent } from './examples/direction';
 
 @Component({
   selector: 'cookbook-label-example',
   templateUrl: './label-example.component.html',
   styleUrls: ['./label-example.component.scss'],
-  imports: [
-    LabelExampleItemComponent,
-    FormFieldTextareaLabelExampleComponent,
-    LabelFormFieldInputLabelExampleComponent,
-    RadioInFormFieldExampleComponent,
-  ],
+  imports: [FormFieldInputLabelExampleComponent, LabelExampleDirectionComponent],
 })
 export class LabelExampleComponent {}
